+++ src/App.vue
@@ -14,9 +14,9 @@
 </template>
 
 <script setup lang="ts">
-import TheHeader from '@/components/TheHeader.vue'
-import TheFooter from '@/components/TheFooter.vue'
-import TheNavigation from '@/components/TheNavigation.vue'
+import TheHeader from '@/components/layout/TheHeader.vue'
+import TheFooter from '@/components/layout/TheFooter.vue'
+import TheNavigation from '@/components/layout/TheNavigation.vue'
 </script>
 
 <style lang="scss">
