+++ src/router/index.ts
@@ -10,11 +10,6 @@ const router = createRouter({
             component: HomeView,
         },
         {
-            path: '/about',
-            name: 'about',
-            component: () => import('../views/AboutView.vue'),
-        },
-        {
             path: '/photos',
             name: 'photos',
             component: () => import('../views/PhotosView.vue'),
